Add files via upload
It includes the following features:
A search bar to find songs
A list of songs to play
Play, pause, and next buttons
A volume slider
A now playing section that displays the currently playing song and album artwork.
The project was created as a way to learn front-end web development and to practice using HTML, CSS, and JavaScript. It is still under development, but it is a good starting point for anyone who wants to create a similar project.